+++ src/components/emptyState/EmptyState.js
@@ -7,7 +7,7 @@ import {
   IllustrationEmptyState48X48Pointer,
   IllustrationEmptyState130X130Pointer,
 } from '@teamleader/ui-illustrations';
-import { Heading3, TextBodyCompact } from '../typography';
+import { Heading3, TextBody } from '../typography';
 import theme from './theme.css';
 
 const illustrationMap = {
@@ -18,7 +18,7 @@ const illustrationMap = {
 
 class EmptyState extends PureComponent {
   render() {
-    const { className, metaText, size, title, ...others } = this.props;
+    const { className, metaText, hidePointer, size, title, ...others } = this.props;
 
     const classNames = cx(
       theme['wrapper'],
@@ -39,13 +39,13 @@ class EmptyState extends PureComponent {
         justifyContent="center"
         textAlign="center"
       >
-        {title && <div className={theme['pointer']}>{illustrationMap[size]}</div>}
+        {title && !hidePointer && <div className={theme['pointer']}>{illustrationMap[size]}</div>}
         <div className={theme['content']}>
           {title && <Heading3 color="teal">{title}</Heading3>}
           {metaText && (
-            <TextBodyCompact marginTop={title && 2} color="neutral">
+            <TextBody marginTop={title && 2} color="neutral">
               {metaText}
-            </TextBodyCompact>
+            </TextBody>
           )}
         </div>
       </Box>
@@ -54,12 +54,14 @@ class EmptyState extends PureComponent {
 }
 
 EmptyState.propTypes = {
+  hidePointer: PropTypes.bool,
   metaText: PropTypes.oneOfType([PropTypes.node, PropTypes.string]),
   size: PropTypes.oneOf(['small', 'medium', 'large']),
   title: PropTypes.oneOfType([PropTypes.node, PropTypes.string]),
 };
 
 EmptyState.defaultProps = {
+  hidePointer: false,
   size: 'medium',
 };