+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ### Added
 
 - `WysiwygEditor` Added `inputClassname` prop. ([@mikeverf](https://github.com/mikeverf) in #1103)
+- `EmptyState`: added `hidePointer` boolean prop (default false). ([@driesd](https://github.com/driesd) in [#1104])
 
 ### Changed
+
+- `EmptyState`: changed meta text component to increase line-height from 18px to 21px. ([@driesd](https://github.com/driesd) in [#1104])
 
 ### Deprecated